Change repositories structure
