+++ app/src/main/java/com/things/smartwateringapp/data/repository/info/InfoRepository.kt
@@ -1,11 +1,9 @@
-package com.things.smartwateringapp.data.repository.home
+package com.things.smartwateringapp.data.repository.info
 
 import com.google.firebase.database.FirebaseDatabase
 import com.things.smartwateringapp.domain.entity.DataInfo
-import com.things.smartwateringapp.domain.entity.Event
 import com.things.smartwateringapp.domain.entity.Status
 import com.things.smartwateringapp.domain.entity.Type
-import durdinapps.rxfirebase2.DataSnapshotMapper
 import durdinapps.rxfirebase2.RxFirebaseDatabase
 import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -13,7 +11,7 @@ import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 
 
-class HomeRepository @Inject constructor(private val database: FirebaseDatabase) {
+class InfoRepository @Inject constructor(private val database: FirebaseDatabase) {
 
     fun getDataInfo(): Single<DataInfo> {
         return RxFirebaseDatabase.observeSingleValueEvent(database.getReference("data"), DataInfo::class.java)
@@ -46,16 +44,5 @@ class HomeRepository @Inject constructor(private val database: FirebaseDatabase)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .toSingle()
-    }
-
-    fun getNearestEvent(): Single<Event> {
-        return RxFirebaseDatabase.observeSingleValueEvent(database.getReference("events"), DataSnapshotMapper.listOf(Event::class.java))
-                .toObservable()
-                .flatMapIterable { it }
-                .filter { it.dateTime > System.currentTimeMillis() }
-                .sorted { event1, event2 -> (event1.dateTime - event2.dateTime).toInt() }
-                .firstOrError()
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
     }
 }
